requirements.txt
ultralytics>=8.0.0
PyYAML
python-dotenv
opencv-python
